project-4/requirements.txt
requests
Pillow
ipython
google-generativeai
python-dotenv